requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
langchain==0.3.0
langchain-community==0.3.0
langgraph==0.2.45
langchain-ollama==0.2.0
langchain-weaviate==0.0.3
weaviate-client==4.9.3
pypdf==5.1.0
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
sentence-transformers==3.3.1
ragas==0.1.20
datasets==2.16.1
